+++ ontimize/layouts/app-layout/o-app-layout.component.ts
@@ -2,12 +2,14 @@ import { Component, NgModule, ViewEncapsulation } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 
+import { Util } from '../../util/util';
 import { OSharedModule } from '../../shared';
 import { InputConverter } from '../../decorators';
 import { OAppHeaderModule } from '../../components/app-header/o-app-header.component';
 import { OAppSidenavModule } from '../../components/app-sidenav/o-app-sidenav.component';
 
 export const DEFAULT_INPUTS_O_APP_LAYOUT = [
+  'mode',
   'sidenavOpened: sidenav-opened',
   'showHeader: show-header',
   'showUserInfo: show-user-info',
@@ -26,8 +28,11 @@ export const DEFAULT_OUTPUTS_O_APP_LAYOUT: any[] = [];
   styleUrls: ['./o-app-layout.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-
 export class OAppLayoutComponent {
+
+  public static defaultMode: string = 'desktop';
+  public static Modes: string[] = ['mobile', 'desktop'];
+
   @InputConverter()
   sidenavOpened: boolean = true;
   @InputConverter()
@@ -36,6 +41,19 @@ export class OAppLayoutComponent {
   showUserInfo: boolean = true;
   @InputConverter()
   useFlagIcons: boolean = false;
+
+  get mode(): string {
+    return OAppLayoutComponent.Modes[this._mode];
+  }
+  set mode(val: string) {
+    let m = OAppLayoutComponent.Modes.find(e => e === val);
+    if (Util.isDefined(m)) {
+      this._mode = m;
+    } else {
+      console.error('Invalid `o-app-layout` mode (' + val + ')');
+    }
+  }
+  protected _mode: string = OAppLayoutComponent.defaultMode;
 
   openedSidenavImg: string;
   closedSidenavImg: string;